GeoLocation double params instead of decimal

+++ GeoChat/GeoChat.Models/GeoLocation.cs
@@ -2,7 +2,7 @@
 {
     public class GeoLocation : Entity
     {
-        public decimal Longitude { get; set; }
-        public decimal Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
     }
 }
